requirements.txt
Flask
Jinja2
requests
numpy
pandas
tqdm
google-generativeai
